[EI-227] [web] refactor: Update import paths for DraggableItem
